trove-backend/trove/requirements.txt
cryptg
django
django-eav2
django_rest_framework
google-generativeai
python-dotenv
telethon
social-auth-app-django